Added config.toml for theme as the colors were acting different for everyone otherwise.
Also updated the prompt and some other small stuff.

We had an issue with the Ollama-model not being able to handle big files, which turned out to be a problem with Ollama itself. It works fine with OpenAI, so we'll let that slide and use smaller files for Ollama instead.